requirements.txt
aioredis==1.3.1
aiosqlite==0.15.0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-braces==1.14.0
django-crispy-forms==1.11.1
django-model-utils==4.1.1
django-private-chat==0.3.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1
djangorestframework-simplejwt==4.6.0
flake8==3.9.0
greenlet==1.0.0
hiredis==1.1.0
httpie==2.4.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
iniconfig==1.1.1
iso8601==0.1.13
mccabe==0.6.1
msgpack==1.0.2
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.2
Pillow==7.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.8.1
PyHamcrest==2.0.2
PyJWT==2.0.1
pymongo==3.11.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPika==0.42.1
PySocks==1.7.1
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-django==4.1.0
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
semantic-version==2.8.5
service-identity==18.1.0
setuptools-rust==0.12.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tortoise-orm==0.16.16
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.58.0
websockets==8.1
whitenoise==6.0.0
zope.interface==5.2.0
